Changed the airline primary key to autoincrement, and changed associated test cases.
This took slightly longer than expected, as now I am retrieving the airline row via the ICAO, so I can set the local airline to the same as the one in the database. However, the ICAOs I have for the fake airlines used in the tests had the same ICAOs as those in the database. Therefore, this was changed.